+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 - Introduced `InstructionFinetuningDataRepository` for storing and retrieving finetuning samples. Comes in two implementations:
   - `PostgresInstructionFinetuningDataRepository` to work with data stored in a Postgres database.
   - `FileInstructionFinetuningDataRepository` to work with data stored in the local file-system.
-
+- Compute precision, recall and f1-score by class in `SingleLabelClassifyAggregationLogic`
 
 ### Fixes
 ...